requirements.txt
coverage==4.0.3
dj-database-url==0.4.0
Django==1.9.2
django-cors-headers==1.1.0
djangorestframework==3.3.2
docker-compose==1.6.0
gunicorn==19.4.5
jinja2==2.8.0
psycopg2==2.6.1
